html changes

+++ client/src/Components/CustomerDetails.jsx
@@ -18,12 +18,13 @@ function GetCustomerDetails() {
     }, [])
 
     const deleteData = async (id) => {
-        await alert("Are you sure?")
-        axios.delete(`${import.meta.env.VITE_SERVER_URL}/delete/${id}`)
-            .then(() =>
-                window.location.reload()
-                )
-            .catch(err => console.log(err))
+        if(confirm("Are you sure?")==true){
+            axios.delete(`${import.meta.env.VITE_SERVER_URL}/delete/${id}`)
+                .then(() =>
+                    window.location.reload()
+                    )
+                .catch(err => console.log(err))
+        }
     }
     if (data === null) {
         <LoadingPage />
